# Conflicts:
#	src/main/java/com/sda/group2/FileInput.java
#	src/main/java/com/sda/group2/Launcher.java
#	src/main/java/com/sda/group2/hibernate/hql/Airport.java
#	src/main/java/com/sda/group2/hibernate/hql/Flight.java

+++ src/main/java/com/sda/group2/hibernate/hql/users/Account.java
@@ -11,7 +11,7 @@ public abstract class Account {
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "account_id", nullable = false)
     private int accountId;
-    @Column (unique = true, nullable = false)
+    @Column(unique = true, nullable = false)
     private String email;
     @Column
     private String password;
@@ -20,13 +20,13 @@ public abstract class Account {
     @Column(name = "last_name", nullable = false)
     private String lastName;
     @Column
-    private int wallet;
+    private double wallet;
 
-    public void setWallet(int wallet) {
+    public void setWallet(double wallet) {
         this.wallet = wallet;
     }
 
-    public int getWallet() {
+    public double getWallet() {
         return wallet;
     }
